Restructure api folder
