+++ api/controller/routes.py
@@ -15,11 +15,11 @@
 #  You should have received a copy of the GNU Affero General Public License
 #  along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-from flask import Response
-from flask_restful import Resource as Restful_Resource
+from .views import *
 
 
-class HealthApi(Restful_Resource):
-    def get(self):
-        # Health check
-        return Response(status=200)
+def initialize_routes(api):
+    api.add_resource(RootApi, '/')
+    api.add_resource(HealthApi, '/health')
+    api.add_resource(ResourcesApi, '/resources')
+    api.add_resource(ResourceApi, '/resources/<resource_id>')
