+++ api/models/resource/CopyrightInfo.py
